Added copy of tensorflow/models/research/audioset
For the intention of not having to force the users to clone or download
the entire repository from https://github.com/tensorflow/models (nearly
500MB), and since only a small subset of the scripts are needed here,
I, the author of this commit, am adding a copy of such files here.

In the process, I cloned the the latest commit at the time
https://github.com/tensorflow/models/commit/f505cecde2d8ebf6fe15f40fb8bc350b2b1ed5dc
Everything was kept from: research/audioset.
The LICENSE statement has been kept intact with the intention to make it
clear that the attribution of the work in the respective scripts belongs
to the appropriate copyright holders.
Everything else was deleted.

I may modify the copied files (except the LICENSE) as per need, but this
commit will be the ultimate source of where I start from.

This is purely for convenience. Please don't sue me.